se separan los hooks useProjectsDatabase y se crea useAddPNewProject
